# doc/api.yaml
openapi: 3.0.0
info:
  title: Knowledge Hub
  description: Knowledge hub service for managing articles, categories, and comments
  version: 1.0.0
servers:
  - url: /api
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        login:
          type: string
          example: 'TestUser'
        role:
          type: string
          enum: [admin, editor, viewer]
          example: 'viewer'
        createdAt:
          type: number
          example: 1655000000
        updatedAt:
          type: number
          example: 1655000000
      required:
        - id
        - login
        - role
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
paths:
  /auth/signup:
    post:
      tags:
        - Auth
      security: []
      summary: Signup
      description: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                login:
                  type: string
                  description: Username
                password:
                  type: string
                  description: Password
              required:
                - login
                - password
      responses:
        201:
          description: Successful signup
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        400:
          description: Bad request
        429:
          description: Too many requests from this IP
  /auth/login:
    post:
      tags:
        - Auth
      security: []
      summary: Login
      description: Login and receive JWT tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                login:
                  type: string
                  description: Username
                password:
                  type: string
                  description: Password
              required:
                - login
                - password
      responses:
        200:
          description: Successful login
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
                  refreshToken:
                    type: string
        403:
          description: Incorrect login or password
        429:
          description: Too many requests from this IP
  /auth/refresh:
    post:
      tags:
        - Auth
      security: []
      summary: Refresh tokens
      description: Get new access and refresh tokens using a valid refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
              required:
                - refreshToken
      responses:
        200:
          description: New token pair
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
                  refreshToken:
                    type: string
        401:
          description: No refresh token provided
        403:
          description: Invalid or expired refresh token
...